Extract generateStaticParams
