+++ src/app/cosmos/[fixture]/nextCosmosRenderer.tsx
@@ -2,7 +2,7 @@ import { RendererConfig, UserModuleWrappers } from 'react-cosmos-core';
 import { NextFixtureLoader } from './NextFixtureLoader';
 import { CosmosPageParams } from './nextShared';
 
-type NextCosmosRendererArgs = {
+type Args = {
   rendererConfig: RendererConfig;
   moduleWrappers: UserModuleWrappers;
 };
@@ -11,10 +11,7 @@ type NextCosmosRendererProps = {
   params: CosmosPageParams;
 };
 
-export function nextCosmosRenderer({
-  rendererConfig,
-  moduleWrappers,
-}: NextCosmosRendererArgs) {
+export function nextCosmosRenderer({ rendererConfig, moduleWrappers }: Args) {
   return function NextCosmosRenderer({ params }: NextCosmosRendererProps) {
     return (
       <NextFixtureLoader
